Support MacOS' version of `date`
Since Macs use the BSD `date` command, the arguments and flags need to
be different. While we're at it, add support for parsing the date string
natively with vim using the builtin `strptime()` function, but only if
it's available, since it was a recent addition.

Neither the Mac `date` command and strptime(), support parsing
fractional seconds, so remove them prior to parsing since they're not
going to affect the result by anything like a meaningful amount.